+++ CHANGELOG.md
@@ -38,6 +38,7 @@ only at the last position
 - Add a `builtins` default feature that gate all filters/functions requiring additional dependencies
 - Add `unique` and `map` filter
 - Add a `timezone` attribute to the `date` filter
+- Add a `get_random` function to get a random number in a range
 
 ## 0.11.20 (2018-11-14)